backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.35.0
sqlalchemy>=2.0.0
alembic>=1.16.0
asyncpg>=0.30.0
python-multipart>=0.0.20
python-jose[cryptography]>=3.5.0
pytest>=8.4.0
pytest-asyncio>=1.0.0
httpx>=0.28.0
python-dotenv>=1.1.0
supabase>=2.16.0
redis>=6.2.0
google-generativeai>=0.8.0
pydantic>=2.11.0